+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Add `implode` method to `ISeq`
 
 ## 3.7.0 - 2018-07-19
 - Allow `mixed`/`any` type for `Generic Collections`